fixedscrolling
So you have not double refs. Now fix insert and cancel.
How?
after finding the ref, verify whether it's the max or not. If it's the max, well, add it simply. If you need to insert, no, you must call scrolling...
With min, you just call scrolling for 1 and after, but with 0 you must do some simple to have a scrolling in [0].
Check riferimentoprimoelemento.
Maybe that's not wrong, as you have to insert a new element and not to scroll.